Prioritizes the kqueue over the poll file descriptor poll class. It change does only apply to MacOS X and BSD systems.
Even though the poll() based back-end was not as fast as the kqueue()
one, it was slightly more stable somehow. Now, I have discovered that
the Darwin poll() implementation does not work properly whenever a
socket buffer (kernel level) is full and its queried from the write
status. It returns the socket as writable when any write() or send()
operation would just fail with an EAGAIN error.

*HEADS UP*: BSD and MacOS X users, this change might introduce some
regressions. In case something went wrong, you could always force the
polling mode to poll() from the 'Advanced' section of cherokee-admin.

git-svn-id: svn://svn.cherokee-project.com/cherokee/trunk@5396 5dc97367-97f1-0310-9951-d761b3857238