This patch changes how log files are handled in Cherokee. Even though it is a very small change, it might have a big impact on may set-ups. The default case scenario is impacted by the change: where the server is launched as root and then re-owned by an unprivileged 'www' user. Log files are now opened *before* the administration role is dropped, so whenever a new log file is created it will be assigned to the root user.
There were a couple of issues with the previous scheme: Firstly, it
did not play along with the default privileges of /var/log on most of
the systems. Second, it was far less secure: since the log file was
writable the 'www' user, an attacker could get to remove log entries
if a security flaw was exploited in the web software.

git-svn-id: svn://svn.cherokee-project.com/cherokee/trunk@5512 5dc97367-97f1-0310-9951-d761b3857238