Detects the Python interpreter version. Warns and disables Cherokee Admin whenever Python 3 is detected.
Due to its numerous incompatibilities, Python 3 will NOT be supported
by Cherokee in the near future. As long as there are operating systems
and Linux distributions where Python 3 is not installed by default (as
Python 2.x is) we will stick to Python 2.x. Once Python 3 becomes the
de-facto Python interpreter, we will translate to Python 3.

(I'd have expected Python folks to have learned the lesson of how PHP
screwed up introducing backward incompatibilities, but it seems I was
wrong. <sigh>)

git-svn-id: svn://svn.cherokee-project.com/cherokee/trunk@6081 5dc97367-97f1-0310-9951-d761b3857238